requirements_test.txt
django<=3.3
django-crispy-forms
sorl-thumbnail
pillow
codecov
factory_boy
pytest
pytest-django
pytest-factoryboy
pytest-cov
selenium

# flake8
# pep8-naming
# mypy
